moved multiple resolvers into the same file
