--- src/main/kotlin/no/skatteetaten/aurora/gobo/resolvers/applicationdeployment/ApplicationDeploymentMutationResolver.kt
@@ -1,21 +1,0 @@
-package no.skatteetaten.aurora.gobo.resolvers.applicationdeployment
-
-import com.coxautodev.graphql.tools.GraphQLMutationResolver
-import graphql.schema.DataFetchingEnvironment
-import no.skatteetaten.aurora.gobo.security.currentUser
-import no.skatteetaten.aurora.gobo.service.ApplicationUpgradeService
-import org.springframework.stereotype.Component
-
-@Component
-class ApplicationDeploymentMutationResolver(
-    private val applicationUpgradeService: ApplicationUpgradeService
-) : GraphQLMutationResolver {
-
-    fun redeployWithVersion(input: ApplicationDeploymentVersionInput, dfe: DataFetchingEnvironment): Boolean {
-        applicationUpgradeService.upgrade(dfe.currentUser().token, input.applicationDeploymentId, input.version)
-        return true
-    }
-
-    fun refreshApplicationDeployment(input: ApplicationDeploymentRefreshInput, dfe: DataFetchingEnvironment) =
-        applicationUpgradeService.refreshApplicationDeployment(dfe.currentUser().token, input.applicationDeploymentId)
-}
