--- src/main/kotlin/no/skatteetaten/aurora/gobo/resolvers/imagerepository/ImageRepositoryQueryResolver.kt
@@ -1,11 +1,0 @@
-package no.skatteetaten.aurora.gobo.resolvers.imagerepository
-
-import com.coxautodev.graphql.tools.GraphQLQueryResolver
-import org.springframework.stereotype.Component
-
-@Component
-class ImageRepositoryQueryResolver : GraphQLQueryResolver {
-
-    fun getImageRepositories(repositories: List<String>) =
-        repositories.map { ImageRepository.fromRepoString(it) }
-}
