+++ src/main/kotlin/no/skatteetaten/aurora/gobo/resolvers/imagerepository/imageRepositoryResolvers.kt
@@ -1,13 +1,24 @@
 package no.skatteetaten.aurora.gobo.resolvers.imagerepository
 
+import com.coxautodev.graphql.tools.GraphQLQueryResolver
 import com.coxautodev.graphql.tools.GraphQLResolver
-import no.skatteetaten.aurora.gobo.resolvers.pageEdges
+import graphql.schema.DataFetchingEnvironment
 import no.skatteetaten.aurora.gobo.integration.imageregistry.ImageRegistryServiceBlocking
 import no.skatteetaten.aurora.gobo.integration.imageregistry.ImageTagType
+import no.skatteetaten.aurora.gobo.resolvers.loader
+import no.skatteetaten.aurora.gobo.resolvers.pageEdges
 import org.springframework.stereotype.Component
 
 @Component
-class ImageRepositoryResolver(val imageRegistryServiceBlocking: ImageRegistryServiceBlocking) : GraphQLResolver<ImageRepository> {
+class ImageRepositoryQueryResolver : GraphQLQueryResolver {
+
+    fun getImageRepositories(repositories: List<String>) =
+        repositories.map { ImageRepository.fromRepoString(it) }
+}
+
+@Component
+class ImageRepositoryResolver(val imageRegistryServiceBlocking: ImageRegistryServiceBlocking) :
+    GraphQLResolver<ImageRepository> {
 
     fun tags(
         imageRepository: ImageRepository,
@@ -29,4 +40,10 @@ class ImageRepositoryResolver(val imageRegistryServiceBlocking: ImageRegistrySer
         val allEdges = matchingTags.map { ImageTagEdge(it) }
         return ImageTagsConnection(pageEdges(allEdges, first, after))
     }
+}
+
+@Component
+class ImageRepositoryTagResolver : GraphQLResolver<ImageTag> {
+
+    fun lastModified(imageTag: ImageTag, dfe: DataFetchingEnvironment) = dfe.loader(ImageTag::class).load(imageTag)
 }
